+++ src/main/java/micdoodle8/mods/galacticraft/core/client/render/entities/RenderEvolvedSpider.java
@@ -90,7 +90,7 @@ public class RenderEvolvedSpider extends RenderLiving
             int j = c0 % 65536;
             int k = c0 / 65536;
             GL11.glDisable(GL11.GL_LIGHTING);
-            OpenGlHelper.setLightmapTextureCoords(OpenGlHelper.lightmapTexUnit, (float)j / 1.0F, (float)k / 1.0F);
+            OpenGlHelper.setLightmapTextureCoords(OpenGlHelper.lightmapTexUnit, j / 1.0F, k / 1.0F);
             GL11.glColor4f(1.0F, 1.0F, 1.0F, 1.0F);
             return 1;
         }
